backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
aiosqlite==0.19.0  # SQLite async adapter

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# PDF Generation and Reporting
weasyprint==60.2
jinja2==3.1.2
reportlab==4.0.7

# File handling and validation
python-magic==0.4.27
pillow==10.1.0

# Email and notifications
aiosmtplib==3.0.1
httpx==0.25.2  # For Slack webhooks

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing API endpoints

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Date and time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Logging
structlog==23.2.0

